some cleanup, enhancement and more things to test and work
