+++ gradle.properties
@@ -24,13 +24,13 @@ curseforgeid=869524
 #########################
 #### Version ranges ####
 #######################
-minecraftrange=[1.16.5,)
-forgerange=[36,)
-forgefmlrange=[36,)
+minecraftrange=[1.18.2,)
+forgerange=[40.2.21,)
+forgefmlrange=[40,)
 neorange=[21,)
 neofmlrange=[3,)
 fabricrange=>=0.14
-javarange=>=8
+javarange=>=17
 
 ##############################
 #### Project information ####
@@ -40,8 +40,8 @@ authors=SrRapero720, Goedix
 authors_list=["SrRapero720", "Goedix"]
 contributors=ZenoArrows, zFERDQFREZrzfq, Kekscussino, cyyynthia
 contributors_list=["ZenoArrows", "zFERDQFREZrzfq", "Kekscussino", "cyyynthia"]
-credits=VideoLAN Team for libVLC, Caprica for made VLCJ, Linus torvalds for make us suffer with Linux
-description=Library with an API using VLC for multimedia integration with Minecraft. Works also standalone
+credits=Powered by FFMPEG and VideoLAN, special thanks to Caprica for vlcj
+description=Library and API providing multimedia processing, source patching, rendering, and math tools.
 
 ####################
 #### Libraries ####
