+++ src/main/java/me/srrapero720/watermedia/tools/IOTool.java
@@ -3,8 +3,6 @@ package me.srrapero720.watermedia.tools;
 import me.srrapero720.watermedia.api.image.decoders.GifDecoder;
 import net.sf.sevenzipjbinding.*;
 import net.sf.sevenzipjbinding.impl.RandomAccessFileInStream;
-import net.sf.sevenzipjbinding.simple.ISimpleInArchive;
-import net.sf.sevenzipjbinding.simple.ISimpleInArchiveItem;
 import org.apache.logging.log4j.Marker;
 import org.apache.logging.log4j.MarkerManager;
 
@@ -15,6 +13,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -79,47 +78,41 @@ public class IOTool {
         return true;
     }
 
-    public static void un7zip(Marker it, Path zipPath) throws IOException { un7zip(it, zipPath, zipPath.getParent()); }
-    public static void un7zip(Marker it, Path zipPath, Path destPath) throws IOException {
-        try (RandomAccessFile file = new RandomAccessFile(zipPath.toFile(), "r");
-             IInArchive archive = SevenZip.openInArchive(null, new RandomAccessFileInStream(file))) {
+    public static void un7zip(Path zipPath) throws IOException { un7zip(zipPath, zipPath.getParent()); }
+    public static void un7zip(Path zipPath, Path destPath) throws IOException {
+        LOGGER.debug(IT, "Unzipping '{}' to directory '{}'", zipPath, destPath);
 
-            ISimpleInArchive simpleInArchive = archive.getSimpleInterface();
-            for (ISimpleInArchiveItem i: simpleInArchive.getArchiveItems()) {
-                Path destination = destPath.resolve(i.getPath());
-                if (i.isFolder()) {
-                    if (!destination.toFile().mkdirs()) {
-                        LOGGER.error(it, "Failed to create directory '{}'", destination.toAbsolutePath().toString());
-                    }
-                    continue;
-                }
-                ExtractOperationResult result;
-
-                result = i.extractSlow(data -> {
-//                    BufferedOutputStream output = new BufferedOutputStream(Files.newOutputStream(destination));
-                    return data.length; // Return amount of consumed data
-                });
-
-                if (result != ExtractOperationResult.OK) {
-                    throw new IOException("Failed to extract file '"+ destination.toAbsolutePath() + "', status code: " + result.name());
+        try (var file = new RandomAccessFile(zipPath.toFile(), "r");
+             var archive = SevenZip.openInArchive(null, new RandomAccessFileInStream(file))
+        ) {
+            final int count = archive.getNumberOfItems();
+            final var extractIds = new ArrayList<Integer>();
+            for (int i = 0; i < count; i++) {
+                if ((boolean) archive.getProperty(i, PropID.IS_FOLDER)) {
+                    File f = destPath.resolve(archive.getProperty(i, PropID.PATH).toString()).toFile();
+                    if (!f.exists() && !f.mkdirs()) throw new IOException("Failed to create directories");
+                } else {
+                    extractIds.add(i);
                 }
             }
+
+            archive.extract(DataTool.unboxArray(extractIds), false, new Seven7ExtractCallback(destPath, archive));
         }
     }
 
     public static void unzip(Path zip) throws IOException { unzip(zip, zip.getParent()); }
-    public static void unzip(Path zip, Path dest) throws IOException {
-        LOGGER.debug(IT, "Unzipping '{}' to directory '{}'", zip, dest);
+    public static void unzip(Path zipPath, Path destPath) throws IOException {
+        LOGGER.debug(IT, "Unzipping '{}' to directory '{}'", zipPath, destPath);
 
-        if (!zip.toString().endsWith(".zip"))
+        if (!zipPath.toString().endsWith(".zip"))
             throw new IOException("Attempted to extract a non .zip file");
-        if (!dest.toFile().exists() && !dest.toFile().mkdirs())
+        if (!destPath.toFile().exists() && !destPath.toFile().mkdirs())
             throw new IOException("Failed to make required directories");
 
-        try (var in = new ZipInputStream(Files.newInputStream(zip))) {
+        try (var in = new ZipInputStream(Files.newInputStream(zipPath))) {
             ZipEntry en = in.getNextEntry();
             while (en != null) { // iterates over entries in the zip file
-                String desPath = dest + File.separator + en.getName();
+                String desPath = destPath + File.separator + en.getName();
                 if (!en.isDirectory()) {
                     unzip$extract(in, desPath); // if the entry is a file, extracts it
                 } else {
@@ -138,5 +131,57 @@ public class IOTool {
             int read;
             while ((read = zipIn.read(bytesIn)) != -1) output.write(bytesIn, 0, read);
         }
+    }
+
+    private static final class Seven7ExtractCallback implements IArchiveExtractCallback {
+        private final IInArchive archive;
+        private final Path destPath;
+        private OutputStream out;
+        private int index = 0;
+
+        public Seven7ExtractCallback(Path destPath, IInArchive archive) {
+            this.archive = archive;
+            this.destPath = destPath;
+        }
+
+        @Override
+        public ISequentialOutStream getStream(int i, ExtractAskMode extractAskMode) throws SevenZipException {
+            this.index = i;
+            try {
+                this.out = Files.newOutputStream(destPath.resolve(archive.getProperty(i, PropID.PATH).toString()));
+                return bytes -> {
+                    try {
+                        out.write(bytes);
+                    } catch (IOException e) {
+                        throw new SevenZipException(e);
+                    }
+                    return bytes.length;
+                };
+            } catch (IOException e) {
+                throw new SevenZipException(e);
+            }
+        }
+
+        @Override
+        public void prepareOperation(ExtractAskMode extractAskMode) {}
+
+        @Override
+        public void setOperationResult(ExtractOperationResult extractOperationResult) throws SevenZipException {
+            if (extractOperationResult != ExtractOperationResult.OK) {
+                LOGGER.error(IT, "Failed to extract file {}", archive.getProperty(index, PropID.PATH));
+            }
+
+            try {
+                out.close();
+            } catch (IOException e) {
+                throw new SevenZipException(e);
+            }
+        }
+
+        @Override
+        public void setTotal(long l) {}
+
+        @Override
+        public void setCompleted(long l) {}
     }
 }
