+++ src/test/java/org/watermedia/network/ImgurPatchTest.java
@@ -1,0 +1,11 @@
+package org.watermedia.network;
+
+import org.junit.jupiter.api.Test;
+
+public class ImgurPatchTest {
+
+    @Test
+    public void testPatch() {
+
+    }
+}
