+++ src/test/java/org/watermedia/render/TestApp.java
@@ -1,0 +1,170 @@
+package org.watermedia.render;
+
+import me.srrapero720.watermedia.WaterMedia;
+import me.srrapero720.watermedia.api.image.ImageAPI;
+import me.srrapero720.watermedia.api.image.ImageCache;
+import me.srrapero720.watermedia.api.image.ImageRenderer;
+import me.srrapero720.watermedia.loader.ILoader;
+import org.lwjgl.glfw.GLFWErrorCallback;
+import org.lwjgl.glfw.GLFWVidMode;
+import org.lwjgl.opengl.ARBDebugOutput;
+import org.lwjgl.opengl.GL;
+import org.lwjgl.system.MemoryStack;
+import org.lwjgl.system.MemoryUtil;
+
+import java.nio.IntBuffer;
+import java.util.LinkedList;
+import java.util.Queue;
+import java.util.concurrent.Executor;
+
+import static org.lwjgl.glfw.Callbacks.glfwFreeCallbacks;
+import static org.lwjgl.glfw.GLFW.*;
+import static org.lwjgl.opengl.GL11.*;
+import static org.lwjgl.system.MemoryStack.stackPush;
+import static org.lwjgl.system.MemoryUtil.NULL;
+
+public class TestApp implements Executor {
+    private Queue<Runnable> executor = new LinkedList<>();
+    private ImageCache cache;
+    private ImageRenderer renderer;
+
+    // The window handle
+    private long window;
+
+    // The media loader
+    private static final String NAME = "org.watermedia.render.TestApp";
+
+    public void run(String url) {
+        try {
+            WaterMedia.prepare(ILoader.DEFAULT).start();
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to load WATERMeDIA", e);
+        }
+
+        renderer = ImageAPI.loadingGif();
+        cache = ImageAPI.getCache(url, this);
+        cache.load();
+
+        init();
+        loop();
+
+        // Free the window callbacks and destroy the window
+        glfwFreeCallbacks(window);
+        glfwDestroyWindow(window);
+
+        // Terminate GLFW and free the error callback
+        glfwTerminate();
+        glfwSetErrorCallback(null).close();
+    }
+
+    private void init() {
+        // Setup an error callback. The default implementation
+        // will print the error message in System.err.
+        GLFWErrorCallback.createPrint(System.err).set();
+
+        // Initialize GLFW. Most GLFW functions will not work before doing this.
+        if ( !glfwInit() )
+            throw new IllegalStateException("Unable to initialize GLFW");
+
+        // Configure GLFW
+        glfwDefaultWindowHints(); // optional, the current window hints are already the default
+        glfwWindowHint(GLFW_VISIBLE, GLFW_FALSE); // the window will stay hidden after creation
+        glfwWindowHint(GLFW_RESIZABLE, GLFW_TRUE); // the window will be resizable
+        glfwWindowHint(GLFW_OPENGL_DEBUG_CONTEXT, GLFW_TRUE);
+
+        // Create the window
+        window = glfwCreateWindow(renderer.width, renderer.height, "WATERMeDIA", NULL, NULL);
+        if ( window == NULL )
+            throw new RuntimeException("Failed to create the GLFW window");
+
+        // Setup a key callback. It will be called every time a key is pressed, repeated or released.
+        glfwSetKeyCallback(window, (window, key, scancode, action, mods) -> {
+            if ( key == GLFW_KEY_ESCAPE && action == GLFW_RELEASE )
+                glfwSetWindowShouldClose(window, true); // We will detect this in the rendering loop
+        });
+
+        // Get the thread stack and push a new frame
+        try ( MemoryStack stack = stackPush() ) {
+            IntBuffer pWidth = stack.mallocInt(1); // int*
+            IntBuffer pHeight = stack.mallocInt(1); // int*
+
+            // Get the window size passed to glfwCreateWindow
+            glfwGetWindowSize(window, pWidth, pHeight);
+
+            // Get the resolution of the primary monitor
+            GLFWVidMode vidmode = glfwGetVideoMode(glfwGetPrimaryMonitor());
+
+            // Center the window
+            glfwSetWindowPos(
+                    window,
+                    (vidmode.width() - pWidth.get(0)) / 2,
+                    (vidmode.height() - pHeight.get(0)) / 2
+            );
+        } // the stack frame is popped automatically
+
+        // Make the OpenGL context current
+        glfwMakeContextCurrent(window);
+        // Enable v-sync
+        glfwSwapInterval(1);
+
+        // Make the window visible
+        glfwShowWindow(window);
+    }
+
+    private void loop() {
+        // This line is critical for LWJGL's interoperation with GLFW's
+        // OpenGL context, or any context that is managed externally.
+        // LWJGL detects the context that is current in the current thread,
+        // creates the GLCapabilities instance and makes the OpenGL
+        // bindings available for use.
+        GL.createCapabilities();
+        ARBDebugOutput.glDebugMessageCallbackARB((source, type, id, severity, length, message, userParam) -> {
+            System.out.println(MemoryUtil.memASCII(message));
+        }, 0);
+
+        // Set the clear color
+        glClearColor(0.0f, 0.0f, 0.0f, 0.0f);
+
+        glEnable(GL_TEXTURE_2D);
+
+        glfwSetWindowSizeCallback(window, (window, width, height) -> {
+            glViewport(0, 0, width, height);
+        });
+
+        // Run the rendering loop until the user has attempted to close
+        // the window or has pressed the ESCAPE key.
+        while ( !glfwWindowShouldClose(window) ) {
+            glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT); // clear the framebuffer
+
+            if (cache.getStatus() == ImageCache.Status.READY)
+                renderer = cache.getRenderer();
+
+            glBindTexture(GL_TEXTURE_2D, renderer.texture(System.currentTimeMillis() % renderer.duration));
+
+            glColor4f(1, 1, 1, 1);
+            glBegin(GL_QUADS);
+                glTexCoord2f(0, 1); glVertex2f(-1, -1);
+                glTexCoord2f(0, 0); glVertex2f(-1, 1);
+                glTexCoord2f(1, 0); glVertex2f(1, 1);
+                glTexCoord2f(1, 1); glVertex2f(1, -1);
+            glEnd();
+
+            glfwSwapBuffers(window); // swap the color buffers
+
+            // Poll for window events. The key callback above will only be
+            // invoked during this call.
+            glfwPollEvents();
+            if (!executor.isEmpty()) executor.remove().run();
+        }
+    }
+
+    public static void main(String[] args) {
+        new TestApp().run(args[0]);
+    }
+
+    @Override
+    public void execute(Runnable command)
+    {
+        executor.add(command);
+    }
+}
